+++ packages/components/molecules/f-rating/src/constants.js
@@ -1,0 +1,8 @@
+// eslint-disable-next-line import/prefer-default-export
+export const VALID_STAR_RATING_SIZES = {
+    small: true,
+    medium: true,
+    large: true
+};
+
+export const VALID_STAR_RATING_DISPLAY_TYPE = ['noRating', 'short', 'medium', 'long'];