+++ packages/components/molecules/f-rating/src/tenants/en-GB.js
@@ -1,7 +1,13 @@
 const messages = {
     locale: 'en-GB',
     ratings: {
-        starsDescription: '{rating} star out of {total} | {rating} stars out of {total}'
+        starsDescription: '{rating} star out of {maxStarRating} | {rating} stars out of {maxStarRating}',
+        ratingDisplayType: {
+            noRating: 'No ratings yet',
+            short: '{reviewCount}',
+            medium: '{rating} of {maxStarRating}',
+            long: 'View {reviewCount} reviews'
+        }
     }
 };