+++ packages/components/molecules/f-rating/src/tenants/es-ES.js
@@ -1,6 +1,12 @@
 export default {
     locale: 'es-ES',
     ratings: {
-        starsDescription: '{rating} ADD TRANSLATION {total} | {rating} ADD TRANSLATION {total}'
+        starsDescription: '{rating} ADD TRANSLATION {total} | {rating} ADD TRANSLATION {total}',
+        ratingDisplayType: {
+            noRating: 'ADD TRANSLATION',
+            short: '{reviewCount}',
+            medium: '{rating} ADD TRANSLATION {maxStarRating}',
+            long: 'ADD TRANSLATION {reviewCount} ADD TRANSLATION'
+        }
     }
 };