+++ packages/components/molecules/f-rating/stories/Rating.stories.js
@@ -12,9 +12,19 @@ export const RatingComponent = (args, { argTypes }) => ({
 
     props: Object.keys(argTypes),
 
-    template: `<rating
-                  v-bind="$props"
-                  :starRating="2" />`
+    template: `
+        <div>
+            <div
+                class="u-spacingBottom--large storybook-grid storybook-grid-columns--4 storybook-grid-stack--lessThanWide"
+                v-for="(list) in ratingVariants">
+                <rating
+                    v-for="(rating) in list"
+                    v-bind="$props"
+                    :starRatingSize="rating.starRatingSize"
+                    :starRating="rating.starRating"
+                    :reviewCount="rating.reviewCount"/>
+            </div>
+        </div>`
 });
 
 RatingComponent.storyName = 'f-rating';
@@ -29,5 +39,32 @@ RatingComponent.argTypes = {
         options: [locales.gb],
         description: 'Choose a locale',
         defaultValue: locales.gb
+    },
+    ratingDisplayType: {
+        control: { type: 'select' },
+        options: ['noRating', 'short', 'medium', 'long', null],
+        description: 'Choose how to display a rating',
+        default: null
     }
+};
+
+RatingComponent.args = {
+    ratingVariants: [
+        [
+            { starRating: 0, starRatingSize: 'small', reviewCount: 499 },
+            { starRating: 2.3, starRatingSize: 'small', reviewCount: 499 },
+            { starRating: 4.5, starRatingSize: 'small', reviewCount: 499 },
+            { starRating: 1.3, starRatingSize: 'small', reviewCount: 499 },
+
+            { starRating: 0, starRatingSize: 'medium', reviewCount: 499 },
+            { starRating: 2.3, starRatingSize: 'medium', reviewCount: 499 },
+            { starRating: 4.5, starRatingSize: 'medium', reviewCount: 499 },
+            { starRating: 1.3, starRatingSize: 'medium', reviewCount: 499 },
+
+            { starRating: 0, starRatingSize: 'large', reviewCount: 499 },
+            { starRating: 2.3, starRatingSize: 'large', reviewCount: 499 },
+            { starRating: 4.5, starRatingSize: 'large', reviewCount: 499 },
+            { starRating: 1.3, starRatingSize: 'large', reviewCount: 499 }
+        ]
+    ]
 };